Исправление проблем базы данных: изменение user_id на BigInteger для поддержки больших Telegram ID

+++ src/models.py
@@ -10,7 +10,7 @@ Base = declarative_base()
 class User(Base):
     __tablename__ = 'users'
 
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     trials_used = Column(Integer, default=0)
     is_paid = Column(Boolean, default=False)
     subscription_expiry = Column(BigInteger, default=0)
@@ -22,7 +22,7 @@ class AuditLog(Base):
     __tablename__ = 'audit_logs'
 
     id = Column(Integer, primary_key=True, autoincrement=True)
-    user_id = Column(Integer, nullable=False, index=True)
+    user_id = Column(BigInteger, nullable=False, index=True)
     event_type = Column(String(100), nullable=False, index=True)
     details = Column(JSON, nullable=False)
     timestamp = Column(DateTime, default=datetime.utcnow, nullable=False, index=True)
